Update scripts revised as per 2.2 release and UC hub documentation

+++ recipes-connectivity/mbed-edge-core/files/rpi3/arm_update_activate.sh
@@ -11,24 +11,35 @@
 #
 . /wigwag/mbed/arm_update_cmdline.sh
 
-echo "-------------------- Executing activate.sh -------------------------"
+set -x
+
+echo "Info: Executing arm_update_activate.sh"
+
+# store the header into partition which survies factory reset
 mkdir -p /userdata/extended
-# copy header to store
-VALUE=$(cp $HEADER /userdata/extended/header.bin)
+if ! cp $HEADER /userdata/extended/header.bin; then 
+    echo "Error: Failed to store the header!"
+    exit 8
+fi
 
+# backup the system and journal logs
 mkdir -p /userdata/.logs-before-upgrade
-cp -R /wigwag/log/* /userdata/.logs-before-upgrade/
+cp -R /var/log/* /userdata/.logs-before-upgrade/
 
-# save version checksum
+# save version file checksum
 md5sum /wigwag/etc/versions.json > /userdata/mbed/version_checksum.md5
 
-systemctl stop maestro
-systemctl start maestro
-
+# save the firmware image to /upgrades partition
+# /upgrades is where the Pelion Edge's upgrade utility expects the image to be
+# checkout out https://github.com/PelionIoT/scripts-pelion-edge to learn more about the update image packaging
 mv $FIRMWARE /upgrades/firmware.tar.gz
-tar -xzf /upgrades/firmware.tar.gz -C /upgrades/
-systemctl start deviceos-wd
-reboot
-echo "-------------------- Finished activate.sh -------------------------"
-exit $VALUE
+if ! tar -xzf /upgrades/firmware.tar.gz -C /upgrades/; then
+    echo "Error: Image copy failed!"
+    exit 5
+fi
 
+# go to middle boot and try applying the update
+reboot
+
+echo "Info: Exiting arm_update_activate.sh"
+exit 0
